requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
python-docx
openpyxl
